Support for LCD and score calculation.
Note that if you change the src/2048_core:score type to uint32_t, then
it would stay 0. The asm seems to be right, so I have no idea who's
fault it is.

+++ include/2048_core.h
@@ -10,10 +10,10 @@
 #ifdef __SDCC
 // TODO: why __pdata would not work.
 typedef __xdata uint8_t board_type[BOARD_SIZE * BOARD_SIZE]; 
-extern __xdata uint32_t score;
+extern __xdata uint16_t score;
 #else
 typedef uint8_t board_type[BOARD_SIZE * BOARD_SIZE]; 
-extern  uint32_t score;
+extern  uint16_t score;
 #endif
 extern board_type board;
